\name{printLanguages}
\alias{printLanguages}
\title{
Convenient user-facing function to print language codes for HP API.
}
\description{
Prints the language codes used by HP IDOL OnDemand's Speech Recognition API. The default in \code{sendAudioGetJobs} is "en-US".
}
\usage{
printLanguages()
}

\examples{
## Prints the language codes used by HP IDOL OnDemand's Speech Recognition API.
printLanguages()
}
